feat: 实现Qdrant向量存储和索引同步功能
本次提交完成了以下核心变更：
1. 新增向量索引抽象层，拆分VectorStore和VectorIndex接口职责
2. 实现Qdrant向量存储后端，支持集合管理、向量写入、检索和文档级删除
3. 新增向量索引同步服务，实现从数据库同步已完成的向量到外部索引
4. 新增文档查询方法，获取可用于索引的已完成向量数据
5. 新增配置项和示例环境变量，支持Qdrant连接配置
6. 重构配置项类型，将qdrant_timeout从float改为int
7. 新增完整的单元测试覆盖向量存储和索引同步逻辑

+++ backend/app/core/config.py
@@ -131,8 +131,8 @@ class Settings(BaseSettings):
         "knowledge_assistant_chunks"
     )
 
-    qdrant_timeout: float = Field(
-        default=10.0,
+    qdrant_timeout: int = Field(
+        default=10,
         gt=0,
     )
